Renamed some things for clarity.
